rename docker_build.rs to commands.rs
