+++ twoliter/src/docker/mod.rs
@@ -1,8 +1,8 @@
-mod docker_build;
+mod commands;
 mod image;
 mod twoliter;
 
-pub(crate) use self::docker_build::DockerBuild;
+pub(crate) use self::commands::DockerBuild;
 pub(crate) use self::image::{ImageArchUri, ImageUri};
 pub(crate) use self::twoliter::create_twoliter_image_if_not_exists;
